requirements.txt
napalm
flask
cryptography.fernet
